Root schema resolver
