+++ Typin/ExitCodes.cs
@@ -22,7 +22,6 @@
         /// <summary>
         /// Gets an exit code from exception.
         /// </summary>
-        [ExcludeFromCodeCoverage]
         public static int FromException(Exception ex)
         {
             return ex switch
